+++ frontend/server/src/Controllers/Problem.php
@@ -39,7 +39,7 @@ namespace OmegaUp\Controllers;
  * @psalm-type ProblemListPayload=array{currentTags: list<string>, loggedIn: bool, pagerItems: list<PageItem>, problems: list<ProblemListItem>, keyword: string, language: string, mode: string, column: string, languages: list<string>, columns: list<string>, modes: list<string>, tagData: list<array{name: null|string}>, tags: list<string>}
  * @psalm-type RunsDiff=array{guid: string, new_score: float|null, new_status: null|string, new_verdict: null|string, old_score: float|null, old_status: null|string, old_verdict: null|string, problemset_id: int|null, username: string}
  * @psalm-type CommitRunsDiff=array<string, list<RunsDiff>>
- * @psalm-type ProblemListCollectionPayload=array{levelTags: list<string>, problemCount: list<array{name: string, problems_per_tag: int}>, otherCollections: list<string>}
+ * @psalm-type ProblemListCollectionPayload=array{levelTags: list<string>, problemCount: list<array{name: string, problems_per_tag: int}>}
  * @psalm-type CollectionDetailsPayload=array{collection: list<array{alias: string, name?: string}>, type: string}
  */
 class Problem extends \OmegaUp\Controllers\Controller {
@@ -4828,8 +4828,7 @@ class Problem extends \OmegaUp\Controllers\Controller {
             'smartyProperties' => [
                 'payload' => [
                     'levelTags' => \OmegaUp\Controllers\Tag::getLevelTags(),
-                    'problemCount' => \OmegaUp\DAO\Problems::getProblemsPerTagCount(),
-                    'otherCollections' => ['problemCollectionAuthors', 'problemCollectionRandomProblem', 'problemCollectionSearchProblem']
+                    'problemCount' => \OmegaUp\DAO\Problems::getProblemsPerTagCount()
                 ],
                 'title' => new \OmegaUp\TranslationString(
                     'omegaupTitleCollections'
