+++ frontend/www/js/omegaup/problem/collection.ts
@@ -15,7 +15,6 @@ OmegaUp.on('ready', () => {
         props: {
           levelTags: payload.levelTags,
           problemCount: payload.problemCount,
-          otherCollections: payload.otherCollections,
         },
       });
     },
